requirements.txt
Django
django-heroku  
gunicorn                                          
psycopg2                                           
pytz                                                  
whitenoise
djangorestframework
docker